requirements.txt
e2b>=1.0.0
groq>=0.9.0
python-dotenv>=1.0.0
httpx>=0.27.0
google-adk>=0.1.0
litellm>=1.0.0
